Finished Collision, Look to desc to see what's next
Need to add a scene/screen where players press a button to join game
(adding a player to the total number of players and having that button
be that player's key). Need to have control work globally. There also
need to have several conditionals add to the planet/sun spawning and
from there we can add the player spawning.